+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Version jump from 0.3->0.7 is to align with other RAPIDS projects.
 - PR #202 Allow insert on start=stop in replace
 - PR #201 Fixed some pad examples in nvstrings.py doc comments
 - PR #186 Fixed memory error in booleans to strings method
+- PR #235 Fix anaconda upload script for new conda recipe
 
 
 # cuStrings/nvStrings 0.3.0 (15 Mar 2019)